Refactor code for better performance and efficiency
I am refactoring this code because the previous one though worked
as expected but experience so much lagging and slow peformance when
changing gridsquares or when sketching that sometimes lead to
freezing.

The above was as a result of creating all the functions in the 'for'
loop used in creating gridsquares couple with the fact that almost
all the functions in the 'for' loop where all inside the
creatInnerDivs' function. I believe this made all the performance
very slow and laggy.

In this new code I will try to have a function for just creating the
grid squares and research on how to target those individual divs so
that each feature will have it's seperate function outside of the
createInnerDivs function.